+++ src/csv.h
@@ -1,2 +1,2 @@
 XMatrix read_x_csv(char *fn, int n, int p);
-double *read_y_csv(char *fn, int n);
+float *read_y_csv(char *fn, int n);
